+++ projects/mission-planner/src/server/features/aeac_scan.py
