+++ projects/mission-planner/src/server/operations/get_info.py
