+++ projects/mission-planner/src/server/operations/takeoff.py
