+++ projects/mission-planner/src/server/utilities/get_autopilot_info.py
