+++ projects/mission-planner/src/server/utilities/wait_for_position_aiding.py
